pyproject.toml
[project]
name = "mcp-app-dynamic-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.1.1",
    "python-fasthtml>=0.12.50",
    "starlette<1.0.0",
]
